# Remove DOCKER_FORCE_BUILD, disable symbolic tags