--- a/.travis.yml
+++ b/.travis.yml
@@ -106,13 +106,14 @@ jobs:
       before_script:
         - gcloud docker --authorize-only
         - bin/docker-pull-deps
+        - bin/docker-pull master
         - |
           export CONDUIT_TAG=$(. bin/_tag.sh ; clean_head_root_tag)
           echo "CONDUIT_TAG=${CONDUIT_TAG}"
         - export PROXY_RELEASE=1 BUILD_DEBUG=1 DOCKER_TRACE=1
 
       script:
-        - bin/docker-build $CONDUIT_TAG
+        - bin/docker-build
 
       after_success:
         - bin/docker-push-deps