--- a/bin/_tag.sh
+++ b/bin/_tag.sh
@@ -14,22 +14,25 @@ go_deps_sha() {
     cat Gopkg.lock Dockerfile-go-deps | shasum - | awk '{print $1}' |cut -c 1-8
 }
 
-dir_tag() {
-    dir="$1"
-    echo "git-$(git log -n 1 --format="%h" "$dir")"
+clean_head() {
+    git diff-index --quiet HEAD --
 }
 
-clean_head_root_tag() {
-    if git diff-index --quiet HEAD -- ; then
-        echo "git-$(git_sha HEAD)"
+head_root_tag() {
+    if clean_head ; then
+        clean_head_root_tag
     else
-        echo "Commit unstaged changes or set an explicit build tag." >&2
-        exit 3
+        echo "dev-$(git_sha HEAD)-$USER"
     fi
 }
 
-master_root_tag() {
-    echo "git-$(git_sha master)"
+clean_head_root_tag() {
+    if clean_head ; then
+        echo "git-$(git_sha HEAD)"
+    else
+        echo "Commit unstaged changes." >&2
+        exit 3
+    fi
 }
 
 validate_tag() {