--- a/bin/docker-build
+++ b/bin/docker-build
@@ -2,20 +2,13 @@
 
 set -eu
 
-. bin/_docker.sh
-. bin/_tag.sh
-
-if [ $# -eq 0 ]; then
-    tag="$(clean_head_root_tag)"
-elif [ $# -eq 1 ]; then
-    tag="${1:-}"
-else
-    echo "usage: $(basename $0) [tag]" >&2
+if [ $# -ne 0 ]; then
+    echo "no arguments allowed for $(basename $0), given: $@" >&2
     exit 64
 fi
 
-bin/docker-build-controller  "$tag"
-bin/docker-build-web         "$tag"
-bin/docker-build-proxy       "$tag"
-bin/docker-build-proxy-init  "$tag"
-bin/docker-build-cli         "$tag"
+bin/docker-build-controller
+bin/docker-build-web
+bin/docker-build-proxy
+bin/docker-build-proxy-init
+bin/docker-build-cli