--- a/bin/docker-build-base
+++ b/bin/docker-build-base
@@ -1,14 +1,20 @@
 #!/bin/sh
 
-# Builds our base runtime docker image.
+# Builds (or pulls) our base runtime docker image.
 
 set -eu
 
+if [ $# -ne 0 ]; then
+    echo "no arguments allowed for $(basename $0), given: $@" >&2
+    exit 64
+fi
+
 . bin/_docker.sh
 
-tag="${1:-2017-10-30.01}"
+tag="2017-10-30.01"
 
-docker_maybe_build . \
-    "$(docker_repo base)" \
-    "${tag}" \
-    Dockerfile-base
+if (docker_pull base "${tag}"); then
+    echo "$(docker_repo base):${tag}"
+else
+    docker_build base "${tag}" Dockerfile-base
+fi