--- a/bin/docker-build-cli
+++ b/bin/docker-build-cli
@@ -2,23 +2,15 @@
 
 set -eu
 
-. bin/_docker.sh
-. bin/_tag.sh
-
-if [ $# -eq 0 ]; then
-    tag="$(clean_head_root_tag)"
-elif [ $# -eq 1 ]; then
-    tag="${1:-}"
-else
-    echo "usage: $(basename $0) [tag]" >&2
+if [ $# -ne 0 ]; then
+    echo "no arguments allowed for $(basename $0), given: $@" >&2
     exit 64
 fi
 
+. bin/_docker.sh
+. bin/_tag.sh
 
 # Build gcr.io/runconduit/cli-bin, which is used by cli/Dockerfile.
-bin/docker-build-cli-bin "${tag}" >/dev/null
+bin/docker-build-cli-bin >/dev/null
 
-docker_maybe_build . \
-    "$(docker_repo cli)" \
-    "${tag}" \
-    cli/Dockerfile
+docker_build cli "$(head_root_tag)" cli/Dockerfile