--- a/bin/docker-build-cli-bin
+++ b/bin/docker-build-cli-bin
@@ -2,36 +2,30 @@
 
 set -eu
 
-. bin/_docker.sh
-. bin/_tag.sh
-
-if [ $# -eq 0 ]; then
-    tag="$(clean_head_root_tag)"
-elif [ $# -eq 1 ]; then
-    tag="${1:-}"
-else
-    echo "usage: $(basename $0) [tag]" >&2
+if [ $# -ne 0 ]; then
+    echo "no arguments allowed for $(basename $0), given: $@" >&2
     exit 64
 fi
+
+. bin/_docker.sh
+. bin/_tag.sh
 
 dockerfile=cli/Dockerfile-bin
 
 validate_go_deps_tag $dockerfile
 
 (
-    unset DOCKER_FORCE_BUILD
     bin/docker-build-base
     bin/docker-build-go-deps
 ) >/dev/null
 
-IMG=$(docker_maybe_build . \
-    "$(docker_repo cli-bin)" \
-    "${tag}" \
-    $dockerfile)
 
+docker_build cli-bin $(head_root_tag) $dockerfile
+IMG=$(docker_repo cli-bin):$(head_root_tag)
 ID=$(docker create "$IMG")
 
-for OS in linux macos ; do
+# copy the newly built conduit cli binaries to the local system
+for OS in darwin linux windows ; do
     DIR="target/cli/${OS}"
     mkdir -p "$DIR"