--- a/bin/docker-build-controller
+++ b/bin/docker-build-controller
@@ -2,29 +2,21 @@
 
 set -eu
 
-. bin/_docker.sh
-. bin/_tag.sh
-
-if [ $# -eq 0 ]; then
-    tag="$(clean_head_root_tag)"
-elif [ $# -eq 1 ]; then
-    tag="${1:-}"
-else
-    echo "usage: $(basename $0) [tag]" >&2
+if [ $# -ne 0 ]; then
+    echo "no arguments allowed for $(basename $0), given: $@" >&2
     exit 64
 fi
+
+. bin/_docker.sh
+. bin/_tag.sh
 
 dockerfile=controller/Dockerfile
 
 validate_go_deps_tag $dockerfile
 
 (
-    unset DOCKER_FORCE_BUILD
     bin/docker-build-base
     bin/docker-build-go-deps
 ) >/dev/null
 
-docker_maybe_build . \
-    "$(docker_repo controller)" \
-    "${tag}" \
-    $dockerfile
+docker_build controller "$(head_root_tag)" $dockerfile