--- a/bin/docker-build-go-deps
+++ b/bin/docker-build-go-deps
@@ -1,13 +1,21 @@
 #!/bin/sh
 
+# Builds (or pulls) our go-deps docker image.
+
 set -eu
+
+if [ $# -ne 0 ]; then
+    echo "no arguments allowed for $(basename $0), given: $@" >&2
+    exit 64
+fi
 
 . bin/_docker.sh
 . bin/_tag.sh
 
-image=$(docker_maybe_build . \
-    "$(docker_repo go-deps)" \
-    "$(go_deps_sha)" \
-    Dockerfile-go-deps)
+tag=$(go_deps_sha)
 
-echo "$image"
+if (docker_pull go-deps "${tag}"); then
+    echo "$(docker_repo go-deps):${tag}"
+else
+    docker_build go-deps "${tag}" Dockerfile-go-deps
+fi