--- a/bin/docker-images
+++ b/bin/docker-images
@@ -5,21 +5,14 @@ set -eu
 . bin/_docker.sh
 . bin/_tag.sh
 
-if [ $# -eq 0 ]; then
-    tag="$(clean_head_root_tag)"
-elif [ $# -eq 1 ]; then
-    tag="${1:-}"
-else
-    echo "usage: $(basename $0) [tag]" >&2
-    exit 64
-fi
-
 docker_image() {
     repo="$(docker_repo "$1")"
     docker image ls \
         --format "{{printf \"%-16s %-10s\" \"$1\" \"${2}\"}} {{.Size | printf \"%6s\"}}  {{.ID}}  {{.CreatedAt}}" \
         "${repo}:${2}"
 }
+
+tag=$(head_root_tag)
 
 docker_image controller   "${tag}"
 docker_image proxy        "${tag}"