--- a/bin/docker-pull
+++ b/bin/docker-pull
@@ -2,14 +2,10 @@
 
 set -eu
 
-. bin/_tag.sh
-
-if [ $# -eq 0 ]; then
-    tag="$(clean_head_root_tag)"
-elif [ $# -eq 1 ]; then
+if [ $# -eq 1 ]; then
     tag="${1:-}"
 else
-    echo "usage: $(basename $0) [tag]" >&2
+    echo "usage: $(basename $0) tag" >&2
     exit 64
 fi