--- a/bin/docker-pull-deps
+++ b/bin/docker-pull-deps
@@ -5,6 +5,6 @@ set -eu
 . bin/_docker.sh
 . bin/_tag.sh
 
-docker_pull base         2017-10-30.01     || true
-docker_pull go-deps      "$(go_deps_sha)"    || true
-docker_pull proxy-deps   "$(proxy_deps_sha)"    || true
+docker_pull base       2017-10-30.01       || true
+docker_pull go-deps    "$(go_deps_sha)"    || true
+docker_pull proxy-deps "$(proxy_deps_sha)" || true