--- a/cli/Dockerfile-bin
+++ b/cli/Dockerfile-bin
@@ -5,8 +5,9 @@ COPY cli cli
 COPY controller controller
 COPY pkg pkg
 RUN mkdir -p /out
+RUN CGO_ENABLED=0 GOOS=darwin go build -a -installsuffix cgo -o /out/conduit-darwin ./cli
 RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o /out/conduit-linux ./cli
-RUN CGO_ENABLED=0 GOOS=darwin go build -a -installsuffix cgo -o /out/conduit-macos ./cli
+RUN CGO_ENABLED=0 GOOS=windows go build -a -installsuffix cgo -o /out/conduit-windows ./cli
 
 ## export without sources & depdenndencies
 FROM gcr.io/runconduit/base:2017-10-30.01