Remove DOCKER_FORCE_BUILD, disable symbolic tags

DOCKER_FORCE_BUILD, combined with symbolic tags, added complexity and
risk of running unintended versions of the code.

This change removes DOCKER_FORCE_BUILD, and sets all Docker tags
programmatically. The decision to pull or build has been moved up the
stack from _docker.sh to the docker-build-* scripts. Workflows that
want to favor docker pulls (like ci), can do so explicitly via
docker-pull.

fixes #141

Signed-off-by: Andrew Seigner <siggy@buoyant.io>

---

### Previous summary for posterity

DOCKER_FORCE_BUILD=1 is a common development build pattern.

Replace (and invert) DOCKER_FORCE_BUILD with DOCKER_TRY_PULL. Default behavior
(with DOCKER_TRY_PULL unset) is to only rely on local docker repo for cached
images. DOCKER_TRY_PULL is enabled in CI. Also remove
DOCKER_BUILD_CACHE_FROM_TAG as it was unused.

fixes #141

# Performance validation

Tested the build before/after code change, with and without force flag, and on empty local docker repo, and subsequent build:

|                 | 1st build | 2nd build |
|-----------------|-----------|-----------|
| no force/before |     3m35s |        8s |
| no force/after  |     3m36s |        8s |
| force/before    |     6m05s |        8s |
| force/after     |     6m03s |        8s |