@olix0r 2018-01-18
> DOCKER_FORCE_BUILD=1 is a common development build pattern.

Can we get a bit clearer on this?

My specific is concern is that the "root" build command, `bin/docker-build` with no other arguments, generally benefits from this behavior.  Should it be more common to run this command with an argument? If so, why?