@siggy 2018-01-18
@olix0r Inverting `DOCKER_FORCE_BUILD` was your suggestion, following a build discussion last Friday, 2018.01.12. Per https://github.com/runconduit/conduit/issues/141#issuecomment-357296715, we discussed two approaches:

Approach 1:
```bash
git commit ...
bin/docker-build && bin/docker-retag-all $(bin/root-tag) latest
```

Approach 2:
```bash
DOCKER_FORCE_BUILD=1 bin/docker-build latest
```

By "common development build pattern" I meant, based on our build discussions last week and today, folks seem to be following Approach 2 (or are not setting `DOCKER_FORCE_BUILD` and unknowingly using docker images that do not contain their code changes). AFAICT, no one (except you) is using Approach 1. I understand that running `bin/docker-build` without arguments benefits from pulling images, but during development that is not a use case that is helpful to solve for, as I want to build an image with my code changes, not pull.

The specific workflow this change is intended to target is documented at: https://github.com/runconduit/conduit/blob/master/BUILD.md#comprehensive.
The goal of this workflow is: edit code, build docker images, run, and repeat, without committing.

The goal of this PR is to slightly simplify the above workflow. Per https://github.com/runconduit/conduit/issues/141#issuecomment-357107109, I'm trying to avoid major retooling of the build, by, for example, modifying how most of the docker-build-* scripts handle arguments. I'm open to going further with this if better defaults/workflows are warranted, but prefer to do this incrementally. Inverting `DOCKER_FORCE_BUILD` decreases the risk that folks aren't running the code they think they are.