@olix0r 2018-01-18
I know that there have been a number of discussions, but for the sake of clarity on issues we still need clear, concrete problem descriptions that don't allude to private meetings.

Also, when I'm discussing anissue and say something like "maybe we should do X?" That's generally an off-the-hip idea and doesn't mean i've particularly thought through it.

I'm not swayed by arguments of "people do x today"-- our dev workflowis still evolving and we should converge on best practices. Why is that the best practice? Help me understand the larger workflow that requires building symbolic tags.

Specifically, I've found tremendous value in using concrete tags (specifically, being sure of what i'm testing -- we've all wasted way too much time testing the wrong images).

Are there impediments to adapting such a workflow? Or do others think this is generally a crazy idea?

If we're going to spend time on this, let's make sure we have a crisp understanding of the problem we're solving and why this is the best (i.e. in terms of tradeoffs) solution for it.