@siggy 2018-01-18
Going to revisit this PR with a new plan:

## Proposed workflow

Goal of this change should be to make edit/build/run as simple as possible:
```
# step 1: build
DOCKER_TRACE=1 bin/mkube bin/docker-build

# step 2: install
# (this may be replaced by a new script)
target/cli/macos/conduit install --version $(bin/root-tag) | kubectl apply -f -
```

Proposed changes:
-