@siggy 2018-01-18
Going to revisit this PR with a new plan:

## Proposed workflow

Goal of this change is to make edit/build/run as simple as possible:
```
# step 1: build
DOCKER_TRACE=1 bin/mkube bin/docker-build

# step 2: install
# (this may be replaced by a new script)
target/cli/macos/conduit install --version $(bin/root-tag) | kubectl apply -f -
```

## Proposed changes

- remove the tag param from the `bin/docker-build` scripts. all tags are concrete, and specified by the tools.
- move implicit docker pulls out of `_docker.sh` up into higher-level scripts that explicitly want to pull, namely:
  - `docker-build-base`
  - `docker-build-go-deps`
  - `docker-build-proxy-deps`
- workflows that want to do docker pulls will do them explicitly, namely, ci